refactor apollo mouse handlers into separate file
